[Bugfix]Switch back to the application on /system when UID and path changed
If the application is a system update app, and the UID of application on /system is changed by OTA, and the version on /system is smaller than the version on /data.
It will remove disabledPkgSetting, and install as a normal app, so need swich back to the application on /system.

Bug: 260683094
Test: manual
Signed-off-by: Tkpointz <pointzztk@gmail.com>