base: Fix smart charging bug
The value of mPowerInputSuspended is switched to true or false in the code, neglecting the fact that
it can also be changed the sysfs node when the phone is plugged/unplugged.

If the user unplugs when the power is above charging limit, mPowerInputSuspended is still true. Therefore,
if it's plugged in again, there will be no attempt to suspend input. The phone will continue to charge.

Fix by reading sysfs node for the actual value whenever updateSmartChargingStatus() is called.

Signed-off-by: Tkpointz <pointzztk@gmail.com>